// src/api/generated/models/QuoteResponse.ts
export interface QuoteResponse {
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    quoteId?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    quoteNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    opportunityId?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    opportunityName?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    customerId?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    customerName?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    contactId?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    contactName?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    quoteStatus?: string;
    /**
     * 
     * @type {Date}
     * @memberof QuoteResponse
     */
    quoteDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof QuoteResponse
     */
    validUntil?: Date;
    /**
     * 
     * @type {number}
     * @memberof QuoteResponse
     */
    subtotal?: number;
    /**
     * 
     * @type {number}
     * @memberof QuoteResponse
     */
    discountAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof QuoteResponse
     */
    taxAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof QuoteResponse
     */
    totalAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    currency?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    paymentTerms?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    deliveryTerms?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    notes?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    assignedUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    assignedUserName?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof QuoteResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof QuoteResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof QuoteResponse
     */
    updatedDate?: Date;
    /**
     * 
     * @type {Array<QuoteItemResponse>}
     * @memberof QuoteResponse
     */
    items?: Array<QuoteItemResponse>;
}